xserver: apply clippy autofix and silence transient nits
Apply `cargo clippy --fix` across the x11-server crate. The
non-generated source warnings collapse via standard idiom fixes:

  - `or_insert_with(Vec::new)` → `or_default()`
  - `if let Some(ref x) = y` → `if let Some(x) = y` where the
    underlying type is Copy
  - `a.len() % 4 != 0` → `!a.len().is_multiple_of(4)` (Rust 1.94
    associated method)
  - `(b - 1) / a + 1` → `b.div_ceil(a)`
  - Unnecessary `u8::from(req.swap_action)` for already-u8 newtypes
  - Unnecessary parens on assigned values
  - Manual `Default` impl reduced to derive where the body was
    `Self::new()` with a parameterless `new`

Two trivial follow-ups: collapse `#[derive(Clone)]\n#[derive(Default)]`
to a single `#[derive(Clone, Default)]` and drop the leftover
parentheses in the colormap RGB-3-3-2 decomposition.

Down to 1 workspace warning (the unfixable `nom v6.1.2`
future-incompat from a transitive dependency). 671 unit tests pass;
38 e2e tests across windows, dbe, randr, and keyboard suites pass.

+++ crates/x11-server/src/xserver/client/sync_flush.rs
@@ -336,8 +336,8 @@ impl ClientState {
                                 damage: damage_id,
                                 timestamp: 0,
                                 area: Rectangle {
-                                    x: x as i16,
-                                    y: y as i16,
+                                    x: x,
+                                    y: y,
                                     width: w,
                                     height: h,
                                 },
